Add files via upload
Review is decided to be luggage related based on ontology from dbpedia
If it contains Damaged: negative reason='damaged luggage'
if it contains lost: negative reason = 'lost luggage'

Similarity score used to test, but accuracy is horrible. So I am going with manual selection of related word of damaged, and lost.